# optimize collecting into `Array` with known element type and length