Fixes #11