@nsajko 2025-05-29
On Julia v1.12 (beta):

* Before:

  ```julia-repl
  julia> using Collects, BenchmarkTools; const c = Collect();

  julia> minimum(@benchmark collect(vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             16.611 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Vector, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             72.477 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Memory, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             34.254 ns
    gctime:           0.000 ns (0.00%)
    memory:           80 bytes
    allocs:           1

  julia> VERSION
  v"1.12.0-beta3"
  ```

* After:

  ```julia-repl
  julia> using Collects, BenchmarkTools; const c = Collect();

  julia> minimum(@benchmark collect(vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             16.471 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Vector, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             23.244 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Memory, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             34.675 ns
    gctime:           0.000 ns (0.00%)
    memory:           80 bytes
    allocs:           1

  julia> VERSION
  v"1.12.0-beta3"
  ```

On Julia v1.13 (nightly):

* Before:

  ```julia-repl
  julia> using Collects, BenchmarkTools; const c = Collect();

  julia> minimum(@benchmark collect(vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             16.702 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Vector, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             67.828 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Memory, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             13.575 ns
    gctime:           0.000 ns (0.00%)
    memory:           80 bytes
    allocs:           1

  julia> VERSION
  v"1.13.0-DEV.655"
  ```

* After:

  ```julia-repl
  julia> using Collects, BenchmarkTools; const c = Collect();

  julia> minimum(@benchmark collect(vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             16.661 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Vector, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             28.815 ns
    gctime:           0.000 ns (0.00%)
    memory:           112 bytes
    allocs:           2

  julia> minimum(@benchmark c(Memory, vec) setup=(vec = rand(Int8, 50);) seconds=Inf samples=100000 evals=1000)
  BenchmarkTools.TrialEstimate:
    time:             12.794 ns
    gctime:           0.000 ns (0.00%)
    memory:           80 bytes
    allocs:           1

  julia> VERSION
  v"1.13.0-DEV.655"
  ```